Configure safe transient agent defaults (Fixes #317)
