+++ src/app_input/tracker_resolver.rs
@@ -137,6 +137,7 @@ mod tests {
             default_agent_kind: jefe::domain::AgentKind::Llxprt,
             transient_agent_dir: std::path::PathBuf::new(),
             default_code_puppy_yolo: None,
+            default_llxprt_mode_flags: Vec::new(),
             transient_max_concurrent: 0,
             agent_ids: Vec::new(),
             default_llxprt_version: None,
