+++ src/domain/mod.rs
@@ -158,6 +158,18 @@ pub struct RemoteRepositorySettings {
 }
 
 /// A repository is a named codebase container.
+fn default_code_puppy_yolo() -> Option<bool> {
+    default_code_puppy_yolo_enabled().then_some(true)
+}
+
+const fn default_code_puppy_yolo_enabled() -> bool {
+    true
+}
+
+fn default_llxprt_mode_flags() -> Vec<String> {
+    vec!["--yolo".to_owned()]
+}
+
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Repository {
     pub id: RepositoryId,
@@ -191,8 +203,11 @@ pub struct Repository {
     #[serde(default)]
     pub transient_agent_dir: PathBuf,
     /// Default Code Puppy YOLO for transient agents. `None` = no yolo.
-    #[serde(default)]
+    #[serde(default = "default_code_puppy_yolo")]
     pub default_code_puppy_yolo: Option<bool>,
+    /// Default LLxprt mode flags for transient agents.
+    #[serde(default = "default_llxprt_mode_flags")]
+    pub default_llxprt_mode_flags: Vec<String>,
     /// Max concurrent transient agents. 0 = no limit (no queueing).
     #[serde(default)]
     pub transient_max_concurrent: u32,
@@ -819,30 +834,50 @@ impl Agent {
         self.origin == AgentOrigin::Transient
     }
 
-    /// Create a transient agent from repository defaults.
+    /// Create a transient agent from an immutable launch snapshot.
+    #[must_use]
+    pub fn new_transient_from_signature(
+        id: AgentId,
+        repository_id: RepositoryId,
+        repo: &Repository,
+        signature: &LaunchSignature,
+    ) -> Self {
+        debug_assert!(
+            signature
+                .work_dir
+                .starts_with(repo.effective_transient_dir()),
+            "transient agent work_dir must be under the repo's effective_transient_dir"
+        );
+        Self {
+            id: id.clone(),
+            display_id: id.0.clone(),
+            repository_id,
+            shortcut_slot: None,
+            name: format!("Transient ({})", repo.name),
+            description: String::new(),
+            work_dir: signature.work_dir.clone(),
+            profile: signature.profile.clone(),
+            code_puppy_model: signature.code_puppy_model.clone(),
+            code_puppy_yolo: signature.code_puppy_yolo,
+            code_puppy_quick_resume: signature.code_puppy_quick_resume,
+            mode_flags: signature.mode_flags.clone(),
+            llxprt_debug: signature.llxprt_debug.clone(),
+            pass_continue: signature.pass_continue,
+            sandbox_enabled: signature.sandbox_enabled,
+            sandbox_engine: signature.sandbox_engine,
+            sandbox_flags: signature.sandbox_flags.clone(),
+            agent_kind: signature.agent_kind,
+            status: AgentStatus::Queued,
+            runtime_binding: None,
+            origin: AgentOrigin::Transient,
+            llxprt_version: signature.llxprt_version.clone(),
+        }
+    }
+
+    /// Create a one-shot transient agent from repository defaults.
     ///
-    /// Transient agents are one-shot: they are created on-the-fly from the
-    /// issue/PR agent chooser, use the repository's default model/options,
-    /// run in a temporary work directory, and are never persisted to
-    /// `state.json`. `pass_continue` is always `false` because transient
-    /// agents are single-session.
-    ///
-    /// # Parameters
-    ///
-    /// - `id` — stable identifier for the new agent.
-    /// - `repository_id` — the repository the transient agent is bound to.
-    /// - `work_dir` — temporary working directory (under the repo's
-    ///   `effective_transient_dir`).
-    /// - `repo` — the source repository whose defaults (profile, model, kind,
-    ///   yolo) are copied.
-    ///
-    /// # Panics (debug only)
-    ///
-    /// In debug builds, panics if `work_dir` is not under the repo's effective
-    /// transient directory. In release builds the check is skipped. Callers
-    /// should always use `generate_transient_work_dir` to produce a valid
-    /// path; this assertion is defense-in-depth against bugs or misuse that
-    /// would escape the expected cleanup/isolation boundary.
+    /// The agent runs under the repository's effective transient directory,
+    /// is never persisted, and cannot continue an earlier session.
     #[must_use]
     pub fn new_transient(
         id: AgentId,
@@ -866,7 +901,7 @@ impl Agent {
             code_puppy_model: repo.default_code_puppy_model.clone(),
             code_puppy_yolo: repo.default_code_puppy_yolo,
             code_puppy_quick_resume: false,
-            mode_flags: Vec::new(),
+            mode_flags: repo.default_llxprt_mode_flags.clone(),
             llxprt_debug: String::new(),
             pass_continue: false,
             sandbox_enabled: false,
@@ -898,7 +933,8 @@ impl Repository {
             issue_base_prompt: String::new(),
             default_agent_kind: AgentKind::default(),
             transient_agent_dir: PathBuf::new(),
-            default_code_puppy_yolo: None,
+            default_code_puppy_yolo: default_code_puppy_yolo(),
+            default_llxprt_mode_flags: default_llxprt_mode_flags(),
             transient_max_concurrent: 0,
             default_llxprt_version: None,
             agent_ids: Vec::new(),
