+++ src/domain/tests.rs
@@ -340,6 +340,7 @@ fn test_issue_base_prompt_serde_roundtrip() {
         default_agent_kind: crate::domain::AgentKind::Llxprt,
         transient_agent_dir: PathBuf::new(),
         default_code_puppy_yolo: None,
+        default_llxprt_mode_flags: Vec::new(),
         transient_max_concurrent: 0,
         agent_ids: vec![],
         default_llxprt_version: None,
@@ -551,7 +552,8 @@ fn repository_new_defaults_transient_fields() {
         PathBuf::from("/tmp/repo"),
     );
     assert!(repo.transient_agent_dir.as_os_str().is_empty());
-    assert_eq!(repo.default_code_puppy_yolo, None);
+    assert_eq!(repo.default_code_puppy_yolo, Some(true));
+    assert_eq!(repo.default_llxprt_mode_flags, vec!["--yolo"]);
     assert_eq!(repo.transient_max_concurrent, 0);
 }
 
@@ -603,6 +605,7 @@ fn agent_new_transient_sets_is_transient_true_and_inherits_repo_defaults() {
     repo.default_profile = "dev".to_string();
     repo.default_code_puppy_model = "gpt-5".to_string();
     repo.default_code_puppy_yolo = Some(true);
+    repo.default_llxprt_mode_flags = vec!["--yolo".to_owned(), "--fast".to_owned()];
     repo.default_agent_kind = AgentKind::CodePuppy;
 
     let work_dir = repo.effective_transient_dir().join("jefe-transient-1");
@@ -620,6 +623,7 @@ fn agent_new_transient_sets_is_transient_true_and_inherits_repo_defaults() {
     assert_eq!(agent.profile, "dev");
     assert_eq!(agent.code_puppy_model, "gpt-5");
     assert_eq!(agent.code_puppy_yolo, Some(true));
+    assert_eq!(agent.mode_flags, vec!["--yolo", "--fast"]);
     assert_eq!(agent.agent_kind, AgentKind::CodePuppy);
     assert!(!agent.pass_continue, "transient agents are one-shot");
     assert_eq!(agent.status, AgentStatus::Queued);
@@ -635,11 +639,12 @@ fn repository_transient_fields_backward_compat_with_missing_fields() {
         "base_dir": "/tmp/repo",
         "default_profile": "",
         "agent_ids": []
-        // Note: no transient_agent_dir, default_code_puppy_yolo, transient_max_concurrent
+        // Note: no transient-agent option fields.
     });
     let repo: Repository = serde_json::from_value(repo_json).value_or_panic("repo serde");
     assert!(repo.transient_agent_dir.as_os_str().is_empty());
-    assert_eq!(repo.default_code_puppy_yolo, None);
+    assert_eq!(repo.default_code_puppy_yolo, Some(true));
+    assert_eq!(repo.default_llxprt_mode_flags, vec!["--yolo"]);
     assert_eq!(repo.transient_max_concurrent, 0);
     assert_eq!(repo.effective_transient_dir(), std::env::temp_dir());
 }
