+++ src/services/tests.rs
@@ -16,6 +16,7 @@ fn local_repository() -> Repository {
         default_agent_kind: crate::domain::AgentKind::Llxprt,
         transient_agent_dir: std::path::PathBuf::new(),
         default_code_puppy_yolo: None,
+        default_llxprt_mode_flags: Vec::new(),
         transient_max_concurrent: 0,
         default_llxprt_version: None,
         agent_ids: Vec::new(),
