+++ src/state/form_ops.rs
@@ -242,6 +242,10 @@ impl AppState {
                     cursor.default_code_puppy_model,
                 );
             }
+            RepositoryFormFocus::DefaultLlxprtMode => {
+                cursor.default_llxprt_mode =
+                    delete_char_before(&mut fields.default_llxprt_mode, cursor.default_llxprt_mode);
+            }
             RepositoryFormFocus::DefaultLlxprtVersion => {
                 cursor.default_llxprt_version = delete_char_before(
                     &mut fields.default_llxprt_version,
@@ -297,6 +301,9 @@ impl AppState {
                     &mut fields.default_code_puppy_model,
                     cursor.default_code_puppy_model,
                 );
+            }
+            RepositoryFormFocus::DefaultLlxprtMode => {
+                delete_char_at(&mut fields.default_llxprt_mode, cursor.default_llxprt_mode);
             }
             RepositoryFormFocus::DefaultLlxprtVersion => {
                 delete_char_at(
