+++ src/state/form_types.rs
@@ -116,6 +116,8 @@ pub struct RepositoryFormFields {
     pub base_dir: String,
     pub default_profile: String,
     pub default_code_puppy_model: String,
+    /// Default LLxprt mode flags for transient agents (form draft).
+    pub default_llxprt_mode: String,
     /// Default LLxprt npm package version (form draft).
     pub default_llxprt_version: String,
     pub default_agent_kind: String,
@@ -149,6 +151,7 @@ pub struct RepositoryFormCursor {
     pub base_dir: usize,
     pub default_profile: usize,
     pub default_code_puppy_model: usize,
+    pub default_llxprt_mode: usize,
     pub default_llxprt_version: usize,
     pub transient_agent_dir: usize,
     pub transient_max_concurrent: usize,
@@ -170,10 +173,11 @@ pub enum RepositoryFormFocus {
     BaseDir,
     DefaultProfile,
     DefaultCodePuppyModel,
-    DefaultLlxprtVersion,
     DefaultAgentKind,
-    TransientAgentDir,
     DefaultCodePuppyYolo,
+    DefaultLlxprtMode,
+    DefaultLlxprtVersion,
+    TransientAgentDir,
     TransientMaxConcurrent,
     GitHubRepo,
     IssuePrRepo,
@@ -196,10 +200,11 @@ impl RepositoryFormFocus {
             Self::BaseDir => Self::DefaultProfile,
             Self::DefaultProfile => Self::DefaultCodePuppyModel,
             Self::DefaultCodePuppyModel => Self::DefaultAgentKind,
-            Self::DefaultAgentKind => Self::DefaultLlxprtVersion,
+            Self::DefaultAgentKind => Self::DefaultCodePuppyYolo,
+            Self::DefaultCodePuppyYolo => Self::DefaultLlxprtMode,
+            Self::DefaultLlxprtMode => Self::DefaultLlxprtVersion,
             Self::DefaultLlxprtVersion => Self::TransientAgentDir,
-            Self::TransientAgentDir => Self::DefaultCodePuppyYolo,
-            Self::DefaultCodePuppyYolo => Self::TransientMaxConcurrent,
+            Self::TransientAgentDir => Self::TransientMaxConcurrent,
             Self::TransientMaxConcurrent => Self::GitHubRepo,
             Self::GitHubRepo => Self::IssuePrRepo,
             Self::IssuePrRepo => Self::RemoteEnabled,
@@ -223,10 +228,11 @@ impl RepositoryFormFocus {
             Self::DefaultProfile => Self::BaseDir,
             Self::DefaultCodePuppyModel => Self::DefaultProfile,
             Self::DefaultAgentKind => Self::DefaultCodePuppyModel,
-            Self::DefaultLlxprtVersion => Self::DefaultAgentKind,
+            Self::DefaultCodePuppyYolo => Self::DefaultAgentKind,
+            Self::DefaultLlxprtMode => Self::DefaultCodePuppyYolo,
+            Self::DefaultLlxprtVersion => Self::DefaultLlxprtMode,
             Self::TransientAgentDir => Self::DefaultLlxprtVersion,
-            Self::DefaultCodePuppyYolo => Self::TransientAgentDir,
-            Self::TransientMaxConcurrent => Self::DefaultCodePuppyYolo,
+            Self::TransientMaxConcurrent => Self::TransientAgentDir,
             Self::GitHubRepo => Self::TransientMaxConcurrent,
             Self::IssuePrRepo => Self::GitHubRepo,
             Self::RemoteEnabled => Self::IssuePrRepo,
