+++ src/ui/screens/new_repository.rs
@@ -153,27 +153,6 @@ pub fn NewRepositoryForm(props: &NewRepositoryFormProps) -> impl Into<AnyElement
             if model_focused { rc.bright } else { rc.fg },
             sel,
         ));
-
-        // Default Code Puppy YOLO for transient agents (issue #213).
-        let yolo_focused = focus == RepositoryFormFocus::DefaultCodePuppyYolo;
-        let yolo_mark = if fields.default_code_puppy_yolo {
-            "x"
-        } else {
-            " "
-        };
-        let yolo_line = format!("  {:<16} [{}]  (space toggles)", "Default YOLO", yolo_mark);
-        all_lines.push(selectable_line(
-            &yolo_line,
-            {
-                let i = line_idx;
-                line_idx += 1;
-                i
-            },
-            selection,
-            pane,
-            if yolo_focused { rc.bright } else { rc.fg },
-            sel,
-        ));
     }
 
     let kind_focused = focus == RepositoryFormFocus::DefaultAgentKind;
@@ -202,6 +181,56 @@ pub fn NewRepositoryForm(props: &NewRepositoryFormProps) -> impl Into<AnyElement
         kind_color,
         sel,
     ));
+
+    if crate::state::is_repository_field_visible(
+        RepositoryFormFocus::DefaultCodePuppyYolo,
+        default_kind,
+    ) {
+        let yolo_focused = focus == RepositoryFormFocus::DefaultCodePuppyYolo;
+        let yolo_mark = if fields.default_code_puppy_yolo {
+            "x"
+        } else {
+            " "
+        };
+        let yolo_line = format!("  {:<16} [{}]  (space toggles)", "Default YOLO", yolo_mark);
+        all_lines.push(selectable_line(
+            &yolo_line,
+            {
+                let i = line_idx;
+                line_idx += 1;
+                i
+            },
+            selection,
+            pane,
+            if yolo_focused { rc.bright } else { rc.fg },
+            sel,
+        ));
+    }
+
+    if crate::state::is_repository_field_visible(
+        RepositoryFormFocus::DefaultLlxprtMode,
+        default_kind,
+    ) {
+        let mode_focused = focus == RepositoryFormFocus::DefaultLlxprtMode;
+        let mode_value = if mode_focused {
+            text_with_caret(&fields.default_llxprt_mode, cursor.default_llxprt_mode)
+        } else {
+            fields.default_llxprt_mode.clone()
+        };
+        let mode_line = format!("  {:<16} [{mode_value}]", "Default Mode");
+        all_lines.push(selectable_line(
+            &mode_line,
+            {
+                let i = line_idx;
+                line_idx += 1;
+                i
+            },
+            selection,
+            pane,
+            if mode_focused { rc.bright } else { rc.fg },
+            sel,
+        ));
+    }
 
     if crate::state::is_repository_field_visible(
         RepositoryFormFocus::DefaultLlxprtVersion,
